Refactor codebase into modular structure
