+++ pyproject.toml
@@ -19,7 +19,9 @@ classifiers = [
 dependencies = [
     "mcp[cli]>=1.6.0",
     "httpx>=0.24.0",
-    "python-dotenv>=1.0.0"
+    "python-dotenv>=1.0.0",
+    "aiofiles>=24.1.0",
+    "tenacity>=9.1.2",
 ]
 
 [project.scripts]
