+++ upstage_mcp/__init__.py
@@ -1,4 +1,8 @@
-"""Upstage Document Parser MCP Server."""
+"""Upstage MCP Server package.
+
+This package provides Model Context Protocol (MCP) integrations for Upstage AI's
+document digitization and information extraction capabilities.
+"""
 
 from .server import main
